backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
redis
PyYAML
